requirements.txt
openpyxl==3.1.5
PyQt5==5.15.11
PyQt5_sip==12.17.0
python_bcrypt==0.3.2
Requests==2.32.4
SQLAlchemy==2.0.41
